+++ CHANGES.md
@@ -6,6 +6,7 @@
 * converting strigui.widget/state :widgets to a hashmap instead of a vector
 * fixing drawing order when widgets are above/below other widgets and 
   how it affects their neighbours
+* fixed tabbing not tabbing through all tabbable widgets
 
 ### 0.0.1-alpha15